dom/tests/mochitest/webcomponents/test_document_shared_registry.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=783129
-->
<head>
  <title>Test shared registry for associated HTML documents.</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<div id="container"></div>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=783129">Bug 783129</a>
<script>
var container = document.getElementById("container");

function registerNoRegistryDoc() {
  var assocDoc = document.implementation.createDocument(null, "html");
  try {
    assocDoc.registerElement("x-dummy", { prototype: Object.create(HTMLElement.prototype) });
    ok(false, "Registring element in document without registry should throw.");
  } catch (ex) {
    ok(true, "Registring element in document without registry should throw.");
  }

  runNextTest();
}

function runNextTest() {
  if (testFunctions.length > 0) {
    var nextTestFunction = testFunctions.shift();
    nextTestFunction();
  }
}

var testFunctions = [
  registerNoRegistryDoc,
  SimpleTest.finish
];

SimpleTest.waitForExplicitFinish();

runNextTest();
</script>
</body>
</html>
